Revert abstract class
